Change of sql id types in cars and user tables Add enumerated annotation in user class Change wrong collumn names in car class

+++ src/main/resources/db/migration/V1.0.0_2022.03.19.07.00__create_users.sql
@@ -7,7 +7,7 @@ create sequence user_id_seq
 
 CREATE TABLE USERS
 (
-    id INT NOT NULL PRIMARY KEY default nextval('user_id_seq'::regclass),
+    id BIGINT NOT NULL PRIMARY KEY default nextval('user_id_seq'::regclass),
     first_name VARCHAR(50),
     last_name VARCHAR(50),
     email VARCHAR(50),
